integration: Enable DHCP in guest image by default
While on host, dnsmasq can be used to autoconfigure the guest network.
However, this requires the guest to be able to pick up the
configuration. The guest agent would usually be responsible for
configuring the network within the guest. A solution might be also
to put a pre-set systemd config, however this should be seen as a
temporary measure.

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>